requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-ibm>=0.3.0
ibm-watsonx-ai>=1.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
scipy>=1.12.0
numpy>=1.26.0
rich>=13.7.0
rank_bm25>=0.2.2
